move slate type definitions to custom.d.ts
