+++ packages/core/admin/admin/custom.d.ts
@@ -1,8 +1,25 @@
-import { StrapiTheme } from '@strapi/design-system';
+import { type StrapiTheme } from '@strapi/design-system';
+import { type Attribute } from '@strapi/types';
+import { type BaseEditor } from 'slate';
+import { type HistoryEditor } from 'slate-history';
+import { type ReactEditor } from 'slate-react';
+
+import { type LinkEditor } from './src/content-manager/components/BlocksEditor/plugins/withLinks';
 
 declare module 'styled-components' {
   // eslint-disable-next-line @typescript-eslint/no-empty-interface
   export interface DefaultTheme extends StrapiTheme {}
+}
+
+declare module 'slate' {
+  interface CustomTypes {
+    Editor: Omit<BaseEditor & ReactEditor & HistoryEditor & LinkEditor, 'children'> & {
+      children: Attribute.BlocksValue;
+    };
+    Element: Attribute.BlocksNode;
+    Descendant: Attribute.BlocksInlineNode | Text;
+    Text: Attribute.BlocksTextNode;
+  }
 }
 
 declare global {
